requirements.txt
tensorflow==2.5.0
numpy==1.19.5
Pillow==8.3.1
opencv-python==4.5.2.54
flask==2.0.1
scipy
